+++ lib/mathjax/mathjax.d.ts
@@ -1,3 +1,4 @@
+import { TeX } from 'mathjax-full/js/input/tex.js';
 import { MathML } from "mathjax-full/js/input/mathml.js";
 import { SVG } from 'mathjax-full/js/output/svg.js';
 import { AsciiMath } from 'mathjax-full/js/input/asciimath.js';
@@ -13,6 +14,7 @@ export declare class MathJaxConfigure {
     mTex: any;
     tex: any;
     texTSV: any;
+    private _validateTex;
     mathjax: any;
     adaptor: any;
     domNode: any;
@@ -25,6 +27,8 @@ export declare class MathJaxConfigure {
     constructor();
     chooseAdaptor: () => void;
     initTex: (nonumbers?: boolean) => void;
+    get validateTex(): TeX<any, any, any>;
+    resetValidateTex: () => void;
     setHandler: (acssistiveMml?: boolean, nonumbers?: boolean) => void;
     changeHandler: (acssistiveMml?: boolean, nonumbers?: boolean) => void;
 }